+++ .github/workflows/integration_test_8gpu_features.yaml
@@ -89,8 +89,8 @@ jobs:
         end=$(date +%s)
         echo "pip install torchao took $((end - start)) seconds"
 
-        # Exercise distributed optimizer tests that are skipped in CPU unit-test CI.
-        python -m pytest tests/unit_tests/test_distributed_muon.py \
+        # Exercise the distributed optimizer check that is skipped in CPU CI.
+        python -m pytest tests/integration_tests/test_distributed_muon.py \
           --durations=20 -vv
 
         sudo mkdir -p "$RUNNER_TEMP/artifacts-to-be-uploaded"
