+++ tests/integration_tests/models.py
@@ -250,9 +250,19 @@ def build_model_tests_list() -> list[OverrideDefinitions]:
                     "--parallelism.data_parallel_shard_degree 4",
                     "--parallelism.expert_parallel_degree 2",
                     "--training.steps 1",
+                    "--checkpoint.enable",
+                    "--checkpoint.enable_first_step_checkpoint",
+                ],
+                [
+                    "--module kimi_k2_7 --config kimi_k2_5_debugmodel",
+                    "--parallelism.spmd_backend spmd_types",
+                    "--parallelism.data_parallel_shard_degree 4",
+                    "--parallelism.expert_parallel_degree 2",
+                    "--training.steps 2",
+                    "--checkpoint.enable",
                 ],
             ],
-            "Kimi K2.7 DistributedMuon spmd_types FSDP+EP",
+            "Kimi K2.7 DistributedMuon spmd_types FSDP+EP checkpoint resume",
             "kimi_k2_5_muon_fsdp+ep_spmd_types",
             ngpu=4,
         ),
